api/requirements.txt
flask
python-dotenv
beautifulsoup4
chromedriver 
selenium
pandas
imgkit
pdfkit
googlesearch-python